move workload_cpu_utilization fixture
