+++ usmqe_tests/conftest.py
@@ -1,5 +1,6 @@
 import configparser
 import pytest
+import datetime
 import usmqe.usmssh as usmssh
 import usmqe.inventory
 
@@ -31,6 +32,27 @@ def get_name(fname):
     # remove 'test_' from the beginning and
     # replace all underscores with spaces
     return fname[5:].replace('_', ' ')
+
+
+def measure_operation(operation):
+    """
+    Get dictionary with keys 'start', 'end' and 'result' that contain
+    information about start and stop time of given function and its result.
+
+    Args:
+        operation (function): Function to be performed.
+
+    Returns:
+        dict: contains information about `start` and `stop` time of given
+            function and its `result`
+    """
+    start_time = datetime.datetime.now()
+    result = operation()
+    end_time = datetime.datetime.now()
+    return {
+        "start": start_time,
+        "end": end_time,
+        "result": result}
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -190,3 +212,31 @@ def os_info():
     config.read_string('[os_info]\n' + f_content)
     LOGGER.debug(config['os_info'])
     return dict(config['os_info'])
+
+
+@pytest.fixture(params=[60, 80, 95])
+def workload_cpu_utilization(request):
+    """
+    Returns:
+        dict: contains information about `start` and `stop` time of stress-ng
+            command and its `result`
+    """
+    def fill_cpu():
+        """
+        Use `stress-ng` tool to stress cpu for 1 minute to given percentage
+        """
+        # stress cpu for for 180 seconds
+        run_time = 180
+        SSH = usmqe.usmssh.get_ssh()
+        host = pytest.config.getini("usm_cluster_member")
+        processors_cmd = "grep -c ^processor /proc/cpuinfo"
+        retcode, processors_count, _ = SSH[host].run(processors_cmd)
+        stress_cmd = "stress-ng --cpu {} -l {} --timeout {}s".format(
+            int(processors_count),
+            request.param,
+            run_time)
+        retcode, stdout, stderr = SSH[host].run(stress_cmd)
+        if retcode != 0:
+            raise OSError(stderr)
+        return request.param
+    return measure_operation(fill_cpu)
